+++ src/main/java/net/frozenblock/trailiertales/worldgen/structure/datagen/GenericRuinsGenerator.java
@@ -51,7 +51,7 @@ public class GenericRuinsGenerator {
 				),
 				RuinsStructure.Type.GENERIC,
 				1F,
-				UniformInt.of(1, 6),
+				UniformInt.of(2, 6),
 				Heightmap.Types.OCEAN_FLOOR_WG
 			)
 		);